Initial commit
This plugin allows to create service users in Gerrit.

A service user is a user that is used by another service to
communicate with Gerrit. E.g. a service user is needed to run the
Gerrit Trigger Plugin in Jenkins. A service user is not able to login
into the Gerrit WebUI and it cannot push commits and tags.

This plugin supports to create service users via SSH, REST and in the
WebUI.

To be able to create a service user a user must be a member of a group
that is granted the 'Create Service User' capability (provided by this
plugin) or the 'Administrate Server' capability.

On creation new service users can be automatically added to configured
groups. This allows to automatically assign or block certain access
rights for the service users.

Change-Id: Ib0106574f7d8f22b50c81364e3437ad3716ba879
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>