Support black list for service user names
Some users tend to create service users with very generic names, such
as 'voter'. Gerrit administrators want to maintain a black list for
service user names to disallow the usage of certain names for service
users. There is a new configuration parameter that allows to specify
usernames that are blocked for service users.

Change-Id: Ie76dc3928022ff71b4b52a6fe8dd66dd035bfaef
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>