Show configurable message on successful service user creation
After the creation of a service user it is helpful to tell the user
what he has to do next, e.g. assign access rights to the service user,
configure the SSH connection on Jenkins so that it can talk to Gerrit
etc. It is now possible to configure an HTML formatted message that is
displayed in the WebUI after a service user was successfully created.

Change-Id: Idb30798dbb1eca4d9b0494548d470ceac7e95ac5
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>