Store for each service user who created it and when it was created
The plugin now stores some properties for the created service users in
'$review_site/data/serviceuser/db', which is a Git config file.

This information may be useful for the Gerrit administrators, e.g. if
a user creates a huge amount of service users the permission to create
service users could be revoked.

In future there may be also other features that require this
information. E.g. a user could list the service users that he created
or he may be allowed to upload additional SSH keys for those users.

Change-Id: I8361897ce04533556624a926df3f493ce2bec91e
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>