Allow enabling of email address for service users
Per configuration it is now possible to define whether service users
are allowed to have an email address or not. Whether an email address
can be assigned for a service user is important because having an
email address e.g. allows the service user to push commits and tags
(if the corresponding access rights are assigned).

This renames the REST endpoints for getting and setting messages since
the REST endpoints are now handling also other configuration
parameters.

Change-Id: I2dcb4acbd05aea50087cd0b9ab31fc59592d8235
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>