Fix checking for 'Create Service User' capability
When using the @RequiresCapability annotation the full capability name
including the plugin name needs to be provided, e.g.
  @RequiresCapability("serviceusers-" + CreateServiceUserCapability.ID)

The plugin name should not be hard coded since it can be changed by
setting Gerrit-PluginName to a different value, hence instead of using
@RequiresCapability we now check the capability programmatically by
using CapabilityControl where we can get the plugin name injected.

Change-Id: I88e4599cd91753a611ccc6c84cc9c83d269b814d
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>