Allow to assign an owner group for a service user
A service user can now have an owner group. If an owner group is
assigned members of this group are able to see and administrate the
service user. If an owner group is assigned the creator of the service
cannot anymore see and administrate the service user if he is not
member of the owner group.

New REST endpoints allow to get/set/delete the owner group and the
owner group is returned as part of the ServiceUserInfo entity.

Change-Id: Ibd7a13ba837a1c35888542adb9cdeab215f06316
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>