Annotate commits of service users by a Git note listing the owners
Commits of a service user can now be annotated by a Git note
that contains information about the current owners of the service
user. This allows to find a real person that is responsible for
this commit. To get such a Git note for each commit of a service
user the 'Forge Comitter' access right must be blocked for service
users.

Change-Id: If8d0c8bcb46fdf1b452b430561c94bba4abef99b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>