Handle users without username
Currently the plugin saves the username of the service user creator in
its database. Some users may not have a username and then the creator
is unknown and trying to lookup the service user creator during the
permission checks fails.

Store the account ID of the service user creator in addition to the
username and do all lookups of the service user creator by using the
account ID instead of the username.

Change ServiceUserInfo so that it now contains an AccountInfo entity
for the creator.

Change-Id: I2aef3bcfb51d2642af32410fe05fcaa62e88022f
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>