Fix setting and deleting of service user email address
Deleting an existing email address failed because the wrong value was
provided to the delete email REST endpoint.

Setting an email address when the user had no email address failed
because an NPE. In this case the REST endpoint to delete the old email
address should not be invoked since there is none.

Change-Id: I5d402dbe91710d7796767aed3377db42dd215a70
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>