Only accept pushes from service users if at least one owner is active
This adds a commit validator that rejects pushes of commits where the
committer is a service user when this service user has no active
owner. The idea of having owners for a service user is that these
owners are responsible for the commits that are created by the service
user. If the accounts of all owners are inactive (e.g. because they
have left the company) the service must not be able anymore to push
further commits because then nobody would be responsible for these
commits.

Change-Id: I15d2130b6c51ff383ccc79d7d842f9846a47c0ab
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>