Fix reloading of configuration changes
When the configuration of the serviceuser plugin is changed (e.g. from
the web UI) then the changes should be effective immediately. This was
broken by Id5819facb591d548df9febf288e38065949128ec which made the
GetConfig REST endpoint singleton. The config object was created in
the constructor and then never updated.

Change-Id: I3f2797983ee5ab690b9806f9b8f4744259598810
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>