Replace icon to clear HTTP passwords with red cross
Gerrit upstream switched from the Tango to the Silk icon set in

  89773d89f928eff3868e590a6584846877b1e249

. With this change, some icons got removed, like the `deleteHover`
icon we used. Gerrit upstream does not offer an icon that symbolizes
"Clear HTTP password", and it's overkill to add additional icons
within this plugin. Hence we use the redNot icon for the time being
(which is a red cross). Once gerrit upstream provides a better icon,
we should switch to that.

Change-Id: I011431eae8dff099702ebe09e04f88eb488ee6b5