Update patch set 6
Patch Set 6:

Thanks, seems to work well!

One small problem though: the plugin creates TopMenu's to create and list service user, which works fine in GWT; however, the topmenus are present in PollyGerrit as well, even though they are not implemented.

There are 2 separate issues here:
- The UI extensions have not been ported yet to PollyGerrit (which is fine, can be done later); but in the mean time the TopMenu's should not be created in PollyGerrit (though they should still be created in GWT)
- The 2 entries are created under the 'People' top menu, which is consistent with GWT; but in PollyGerrit this has been removed, so they should be created elsewhere : probably have a single menu 'Service User' under the 'Browse' top menu, with a button on that page.

On master this will not be a problem, but do you know there is a way to 'disable' the top menus on 2.15/2.16 only when PollyGerrit is used, i.e. an API that may be called from the TopMenu's getEntries() method?

Patch-set: 6
Reviewer: Gerrit User 1017568 <1017568@173816e5-2b9a-37c3-8a2e-48639d4f1153>