Remove GWT-based UI
GWT-based UIs are not supported anymore with Gerrit 3.0. Thus it was not
possible to use the plugin with newer Gerrit versions.

This change removes all components used to create the UI and thus the
dependency to GWT. The aim is to reimplement a new UI using PolyGerrit
in a later change.

Change-Id: I672a4e80f0fcf50ff7156bfae4b800a32b6ec398