Add PolyGerrit UI to list service users
The GWT-based UI was removed recently to allow compatibility with Gerrit
3.0+.

This change adds a screen listing all service users as a first step to
reimplement the UI with PolyGerrit. The new screen can be found under
'Browse' > 'Service Users'.

The list does currently not support pagination, since the REST API of
the plugin does not support it yet. Also, filtering is currently not
implemented, since the router does not support query operators for
plugin screens.

Change-Id: Ibcbfac311e9a279d14dc92299e6d963cef19a87b