Add PolyGerrit UI for showing service user details
This change adds another screen, that lists the account details of
a service user. This screen can be accessed by clicking on the username
in the service user list or by using the URI
/x/serviceuser/user/${account_id}.

The page currently does not allow to edit the service user's account
details. This will be added in a future change.

Change-Id: I6e7e9feeb0d1778e87688ccfaccb631d76045870