Add validation for SSH-key before account creation
If an invalid SSH-key was used, when creating a service user, the account
creation could fail. However, an account was created, which was in the
following not registered as a service user, basically remaining unused
and blocking usernames.

This change adds a validation, checking for the common beginnings of
public keys created by OpenSSH or following RFC standards.

Change-Id: Ifdc7b74a5b723c2e2b913c6eecf70accd09b88b3