Allow additional SSH-key formats
So far only RSA-type SSH-keys were allowed to be used for service users.
Since Gerrit and older versions of this plugin allowed other formats
as well, this could break existing installations.

This change adds a validator-class for SSH-keys that is able to validate
a wider range of SSH-key formats. The validation is far from perfect and
only meant to minimize the risk of creating unusable accounts, because
of an invalid SSH-key.

Change-Id: I3396ee449f6f88708514583a055a45c69abb4dd9