Make full name and email in details screen better readable
The full name and email inputs in the serviceuser detail screen were
not well readable, if they were longer than the input width. In addition
the values could not be selected and copied.

Now the inputs are wider to increase the amount of readable test. Also
the default value is selectable and can thus be copied.

Change-Id: I2653c7ef84e62c44ae3d0506f2f3c79500e87908