Check if full name and email edit are allowed in realm
So far it was not checked, whether it was allowed to change the full
name or the email in a configured auth realm. This could lead to errors,
if a user tried to do it anyway.

This change checks the server config, if the edit of those fields is
allowed, and if not just shows the set value instead of an input.

Change-Id: I14950fe0daf8540be27e10dfafe5b7e5f0a12ff2