Make compatible with Gerrit 3.3.0
In https://gerrit-review.googlesource.com/c/gerrit/+/275992,
`ProjectLevelConfig`'s inheritence was changed to not extend
`VersionedMetaData`. As a result it no longer has a `.commit()` method
on it.

Instead, read the config directly from the git repo and manipulate it
there. This probably makes some subtle cache bugs go away as a side
effect, but it likely isn't important since it would only matter in a
high contention service user creation situation. Such a situation is
difficult to imagine.

Change-Id: I4c7c94dd4db189c4a5e818ac184ddb2c318ebcb4