Allow blocking of usernames using wildcards and regex
So far, only usernames matching exactly a configured string could be
blocked from being used for a service user. This could be tedious.

Now, next to exact matching, a wildcard at the end of the block
definition can be used, as well as regex, if the pattern string
starts with '^'.

Change-Id: Idd668efc1fccf2dce009d1ebaca7079486b7c4ea