Ensure in-memory model matches state committed to All-Projects
Evict the All-Projects project cache entry after committing changes
to All-Projects in order to ensure the in-memory model and committed
state always match.

If commit() fails we observed that the in-memory model and the commit
state may diverge which has the effect that a user will see a new
service user entry in the UI which disappears when the service user
plugin is reloaded or the All-Projects project cache entry is evicted
for another reason.

Change-Id: Iab01a178bd2e9e90fc491f7c1132c161f39dc631