Add option to allow to set custom HTTP passwords
It was only possible for administrators to set custom passwords for
serviceusers. However, there might be a legitimate reason for owners
of a serviceuser to do so, instead of only generating a password.

This change adds an option that allows serviceuser owners to set a
custom HTTP password. This was made optional, since allowing this
might not be compliant with all security standards.

Change-Id: Id674fcbe312f7d912790e0b7fecaa9fd093a085b