Allow serviceuser owners to delete SSH keys
Owners of a serviceuser were not able to delete the
serviceuser's SSH key, because in Gerrit core except for
administrators only deletion of one's own SSH keys is
permitted.

Instead of reusing the DeleteSshKey.apply() method from
Gerrit core, it is now reimplemented in the plugin.
Checking the permissions was removed, since ownership
is already checked, when checking the visibility of the
serviceuser. Email-notifications were removed as well.

Change-Id: Ic530c26ded47c2da096cb8487a9217ba23c1bff0