Add allowSsh option to allow owner to modify SSH keys
Currently everyone can manage the SSH keys of service users, this is of
course undesired behavior and probably a security bug.

Only the owner and admin should be able to modify the SSH keys, in a
similar manner to all the other owner managed attributes.

The allowSsh option was added to match the behaviour of all owner
enabled attributes and a matching access check was added.

The DeleteSshKey is blocked by Gerrit access control as it assumes only
admin and self can delete keys. A change for Gerrit will be submitted.

Change-Id: Ibf9475a9b2ebbb3544ed2c729e4ea4edcf7f4403
Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>