Add allowSsh option to allow owner to modify SSH keys
Currently owners can manage the SSH keys of service users, this differs
from the way that owner can manage other types of credentials and
attributes such as HTTP credentials.

The admin and service creator should be able to control what owner can
set including the SSH keys.

The allowSsh option was added to match the behaviour of all owner
enabled attributes and a matching access check was added.

The DeleteSshKey is blocked by Gerrit access control as it assumes only
admin and self can delete keys. A change for Gerrit will be submitted.

NOTICE: This is a breaking change, previous implementation allowed the
service owner to modify SSH public key, this patch adds the allowSsh
restriction so that administrator must explicitly enable the
functionality.

Change-Id: I232b7536604e4be63ac6fdfffe8bc156f53b9d66
Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>