Fix another link for context path users
When making the plugin work with context paths I missed
that there was a link to the service users list from the
left hand side menu when browsing repo/groups, so I missed
to fix that part. This is now remedied.

I could not as easily get the context path from the
plugin.ts file so I changed the approach. For consistency
I changed the other parts as well.

Solves: bug 350779993
Release-Notes: Make serviceuser plugin work with context paths
Change-Id: I6dfd2de529932c59c28aadea485de5195bcbeff3